refactor: wasm bindings
