+++ x/amm/client/wasm/query_swap_estimation.go
@@ -1,0 +1,11 @@
+package wasm
+
+import (
+	wasmvmtypes "github.com/CosmWasm/wasmvm/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	wasmbindingstypes "github.com/elys-network/elys/wasmbindings/types"
+)
+
+func (oq *Querier) querySwapEstimation(ctx sdk.Context, query *wasmbindingstypes.QuerySwapEstimationRequest) ([]byte, error) {
+	return nil, wasmvmtypes.UnsupportedRequest{Kind: "QuerySwapEstimation, not implemented yet"}
+}
