+++ x/oracle/client/wasm/messenger.go
@@ -1,0 +1,27 @@
+package wasm
+
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	wasmbindingstypes "github.com/elys-network/elys/wasmbindings/types"
+	"github.com/elys-network/elys/x/oracle/keeper"
+)
+
+// Messenger handles messages for the Oracle module.
+type Messenger struct {
+	keeper *keeper.Keeper
+}
+
+func NewMessenger(keeper *keeper.Keeper) *Messenger {
+	return &Messenger{keeper: keeper}
+}
+
+func (m *Messenger) HandleMsg(ctx sdk.Context, contractAddr sdk.AccAddress, contractIBCPortID string, msg wasmbindingstypes.ElysMsg) ([]sdk.Event, [][]byte, error) {
+	switch {
+	// case *MsgPriceFeed:
+	//     // Handle the MsgPriceFeed message
+	//     // ...
+	default:
+		// This handler cannot handle the message
+		return nil, nil, wasmbindingstypes.ErrCannotHandleMsg
+	}
+}
